[CBRD-24242] If is_fetch_completed of the EXECUTE protocol is 1, the fetch request is not made. (#43)
http://jira.cubrid.org/browse/CBRD-24242

When using ODBC Gateway, ODBC does not provide the row count of query results, so the total row count cannot be sent to the driver. In this case, an error occurs in the Driver.
However, CAS sends is_fetch_completed set to 1 to the driver when it can no longer send query results.
Although the driver makes a request to CAS as many as the total row count, fetch can be terminated when is_fetch_completed is 1 even if the total rows count is unknown.